neo_dolfin/ai/chatbot_api/requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
chromadb
pypdf
python-dotenv
huggingface_hub
transformers
torch
termcolor
textblob
SpeechRecognition
sentence-transformers
